# integrations/csm-integration-service/openapi.yaml
openapi: 3.0.1
info:
  title: CSM Integration Service API
  version: 1.0.0
  description: >
    Project/account search and their Contacts sub-resource, incident
    creation, and alert-incident mapping create/lookup, for third-party
    (M2M) consumers. Fronted by Choreo's API Manager gateway (subscription +
    client credentials) — this service performs no inbound auth of its own.
security:
  - oauth2ClientCredentials: []
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /cases/{id}:
    patch:
      summary: Update a case's state, severity, or workState (or other fields on ServiceNow).
      description: >
        The request body is forwarded to the entity service verbatim; the
        entity service enforces its own field-combination rules (e.g. exactly
        one of state/severity/workState) and 400s otherwise. Unlike
        PATCH /projects/{id}, whether this operation requires a forwarded
        end-user identity token depends on entity-service's own data source.
        On a Postgres data source, a state/severity/workState update (optionally
        combined with resolutionCode/cause/closeNotes when state is closed or
        solution_proposed) succeeds through this M2M-only service — no token
        check on that path. Every other field this shape accepts (watchList,
        assigneeEmail, type and its companions, parentId, relatedCaseId,
        autocloseHoldUntil, subject, description, deploymentId,
        deployedProductId, the fix-ETA group, acknowledge, workaroundProvided)
        is Postgres-rejected with 400, not proxied through to a token check.
        On a ServiceNow data source, every field this operation accepts —
        including a bare state/severity/workState update — requires a
        forwarded identity token this service cannot supply, so every call
        currently receives a mapped 401 there, with no field combination that
        succeeds.
      operationId: patchCase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Case fields to update. Field-combination rules are enforced upstream.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseUpdateRequest'
      responses:
        "200":
          description: The case was updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseUpdateResponse'
        "400":
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized (upstream reported a missing or invalid forwarded identity — always when entity-service runs a ServiceNow data source, given this service's M2M-only design; not returned on a Postgres data source, where an unsupported field is a 400 instead)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "413":
          description: RequestEntityTooLarge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cases/{id}/comments:
    post:
      summary: Add a comment to a case.
      description: >
        The caller supplies only "type" and "content"; this service supplies
        entity-service's actorEmail field itself, from its own configured
        trusted M2M identity (UMT_INTEGRATION_ACTOR_EMAIL), the same way it
        does for POST /cases/{id}/tags — never from the caller, since
        an unchecked caller-supplied actorEmail would let any caller claim to
        be any user. On DATA_SOURCE=postgres, that configured value must
        match an entry in entity-service's M2M_TRUSTED_ACTOR_EMAILS allowlist
        or every call here receives a mapped 403; when it matches, the
        comment is created successfully with no forwarded end-user identity
        needed. On DATA_SOURCE=servicenow, entity-service's ServiceNow path
        still requires a forwarded end-user identity token unconditionally
        and ignores actorEmail entirely — since this service is strictly M2M
        with no mechanism to supply one, calls against that data source
        still receive a mapped 401.
      operationId: createCaseComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCommentCreateRequest'
      responses:
        "201":
          description: The comment was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseCommentCreateResponse'
        "400":
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: >
            Unauthorized — always for DATA_SOURCE=servicenow, since this
            service has no forwarded end-user identity to supply and that
            data source's entity-service path requires one unconditionally.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: >
            Forbidden — the configured UMT_INTEGRATION_ACTOR_EMAIL is unset
            or is not on entity-service's M2M_TRUSTED_ACTOR_EMAILS allowlist
            (DATA_SOURCE=postgres only).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "413":
          description: RequestEntityTooLarge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          # Placeholder — fill in with the actual token endpoint for this service's
          # Choreo-managed API once it's provisioned. Not a real/guessed URL.
          tokenUrl: "<token_url>"
          scopes: {}
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...
    CaseUpdateRequest:
      type: object
      minProperties: 1
      description: >
        state, severity, and workState are mutually exclusive of each other
        and of every other field in this request — exactly one of them, or
        one combination of the other (ServiceNow-only) fields, is required
        per call. Only state/severity/workState succeed through this
        M2M-only service; every other field requires a forwarded end-user
        identity token this service cannot supply. See the entity service's
        own OpenAPI spec for the full field set and combination rules — this
        is deliberately not fully re-specified here.
      properties:
        state:
          type: string
        severity:
          type: string
        workState:
          type: string
        markFixIssued:
          type: boolean
          description: >
            True-only, first-write-wins: records that a fix has been issued
            for the case. Mutually exclusive with state, severity, workState,
            and every other field in this request. A second call on an
            already-marked case is a no-op success, not an error. Postgres
            data source only (or postgres-servicenow-dual-write, which
            asynchronously mirrors the write to ServiceNow) — unlike
            state/severity/workState, this succeeds through this M2M-only
            service unconditionally, with no data-source-dependent 401.
          enum: [true]
      additionalProperties: true
    CaseUpdateResult:
      type: object
      properties:
        id:
          type: string
        updatedOn:
          type: string
          format: date-time
        updatedBy:
          type: string
        state:
          type: string
        severity:
          type: string
        workState:
          type: string
          nullable: true
      additionalProperties: true
    CaseUpdateResponse:
      type: object
      properties:
        message:
          type: string
        case:
          $ref: '#/components/schemas/CaseUpdateResult'
    CaseCommentCreateRequest:
      type: object
      required:
        - type
        - content
      properties:
        type:
          type: string
          description: Comment type, e.g. comment, work_note, or activity.
        content:
          type: string
          description: >
            actorEmail is never accepted here — this service supplies it
            itself from its own configured M2M identity
            (UMT_INTEGRATION_ACTOR_EMAIL).
    CaseCommentCreateResult:
      type: object
      properties:
        id:
          type: string
        createdOn:
          type: string
          format: date-time
        createdBy:
          type: string
    CaseCommentCreateResponse:
      type: object
      properties:
        message:
          type: string
        comment:
          $ref: '#/components/schemas/CaseCommentCreateResult'
...